pwil/requirements.txt
dm-acme>=0.1.4
mujoco-py>=2.0.2.10
numpy>=1.18.5
gym>=0.17.2
dm-sonnet>=2.0.0
tensorflow-probability>=0.10.0
trfl>=1.1.0
pot>=0.7.0
scikit-learn>=0.23.1
tf-nightly>=2.3.0.dev
dm-reverb-nightly>=0.1.0.dev
